move BusyAnimation to TouchStyle
