+++ board/fischertechnik/TXT/rootfs/opt/ftc/TouchStyle.py
@@ -6,10 +6,14 @@
 import struct, os, platform, socket
 
 from PyQt5.QtCore import *
+from PyQt5.QtCore import pyqtSignal, QPoint, QTimer, Qt
 from PyQt5.QtGui import *
+from PyQt5.QtGui import QColor, QImage, QPainter, QBrush
 from PyQt5.QtWidgets import *
 
 __version__ = '1.8'
+
+from PyQt5.QtWidgets import QWidget
 
 TouchStyle_version = float(__version__)  # Kept for backward compatibility
 
@@ -496,3 +500,77 @@ class TouchApplication(QApplication):
         import touch_keyboard
         self.installEventFilter(touch_keyboard.TouchHandler(self))
         TouchSetStyle(self)
+
+
+class BusyAnimation(QWidget):
+    """Shows a busy animation on top of any application.
+    """
+    expired = pyqtSignal()
+    BUSY_TIMEOUT = 20
+
+    def __init__(self, app, parent=None):
+        super(BusyAnimation, self).__init__(parent)
+        self.resize(64, 64)
+        # center relative to parent
+        self.move(QPoint(parent.width() // 2 - 32, parent.height() // 2 - 32))
+        self.step = 0
+        self.app = app
+        # create a timer to close this window after 10 seconds at most
+        self.etimer = QTimer(self)
+        self.etimer.setSingleShot(True)
+        self.etimer.timeout.connect(self.timer_expired)
+        self.etimer.start(BUSY_TIMEOUT * 1000)
+        # animate at 5 frames/sec
+        self.atimer = QTimer(self)
+        self.atimer.timeout.connect(self.animate)
+        self.atimer.start(200)
+        # create small circle bitmaps for animation
+        self.dark = self.draw(16, QColor("#808080"))
+        self.bright = self.draw(16, QColor("#fcce04"))
+
+    def draw(self, size, color):
+        img = QImage(size, size, QImage.Format_ARGB32)
+        img.fill(Qt.transparent)
+        painter = QPainter(img)
+        painter.setPen(Qt.white)
+        painter.setRenderHint(QPainter.Antialiasing, True)
+        painter.setBrush(QBrush(color))
+        painter.drawEllipse(0, 0, img.width() - 1, img.height() - 1)
+        painter.end()
+        return img
+
+    def timer_expired(self):
+        # App launch expired without callback ...
+        self.expired.emit()
+
+    def animate(self):
+        # if the app isn't running anymore then stop the
+        # animation. This typically only happens for non-txt-styled
+        # apps or with apps crashing.
+        #
+        # We might tell the user that the app ended unexpectedly
+        if self.app.poll():
+            self.close()
+            return
+        self.step += 45
+        self.repaint()
+
+    def close(self):
+        self.etimer.stop()
+        self.atimer.stop()
+        super(BusyAnimation, self).close()
+        super(BusyAnimation, self).deleteLater()
+
+    def paintEvent(self, event):
+        radius = min(self.width(), self.height()) // 2 - 16
+        painter = QPainter()
+        painter.begin(self)
+        painter.setRenderHint(QPainter.Antialiasing)
+        painter.translate(self.width() // 2, self.height() // 2)
+        painter.rotate(45)
+        painter.rotate(self.step)
+        painter.drawImage(0, radius, self.bright)
+        for i in range(7):
+            painter.rotate(45)
+            painter.drawImage(0, radius, self.dark)
+        painter.end()
